Added Teams, and Basic Ready up logic
Every command now checks for teams and returns the teams designated number

The server now uses the team numbers to determine if a game is ready to start,
then launches the new Start_Match() function

On a side note, Handle_Commands() is getting a bit too big and ugly for my taste.
I might move functionality to different functions.