Added a few new regexes for future use
Just setting up most of the command regexes now to get it out of the way.

Most notable absence, teamname_{1,2}, because that will be a little harder.

Also .start can for now only use the current map and will need to take
multiple arguments in the future.

Simple commit because I had a little time before class