Updated Config File
The config file has been given a second look and is now provided as an
example file that needs to be renamed or copied to work. Currently
only "config.toml" will work but it shouldn't be too hard to make it a
command line argument.

Also I removed the unnecessary fullAddr variable from the ServerInfo struct.